// src/app/client/model/personData.ts
/**
 * ConApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PersonData { 
    userId?: string;
    /**
     * Name on birth, never changes
     */
    name?: string | null;
    birthday?: string;
    birthPlace?: string | null;
    /**
     * Category of this datapoint
     */
    category?: string | null;
    /**
     * Field name
     */
    key?: string | null;
    /**
     * The content of this datapoint
     */
    value?: string | null;
    /**
     * When this datapoint was last updated
     */
    changedAt?: string;
}