// src/app/client/model/searchResult.ts
/**
 * ConApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SearchResult { 
    name?: string | null;
    description?: string | null;
    id?: string | null;
    link?: string | null;
}